refactor: split services, add pages/styles dirs, fix core bugs #81
- Split application/services.rs into 5 service files
- Add thumb_path to Rust Asset struct and frontend type
- Fix AssetDetailPanel note loading via get_asset_note API
- Add scan button to Toolbar with loading state
- Implement folder selection in AddLibraryDialog/MoveAssetsDialog
- Add React.memo to AssetGridItem
- Fix thumbnail display using thumb_path when ready
- Split CSS into tokens/variants/base modules
- Fix toast dark mode colors
- Fix scan_library DB lock deadlock risk
- Add selectFolder utility to API client
- Add getAssetNote API function
