+++ src-tauri/src/application/mod.rs
@@ -1,3 +1,11 @@
-pub mod services;
+pub mod library_service;
+pub mod asset_service;
+pub mod tag_service;
+pub mod post_service;
+pub mod move_service;
 
-pub use services::*;
+pub use library_service::LibraryService;
+pub use asset_service::AssetService;
+pub use tag_service::TagService;
+pub use post_service::PostService;
+pub use move_service::{MoveService, MoveResult};
