+++ src-tauri/src/commands/library.rs
@@ -48,13 +48,31 @@ pub fn scan_library(
     let job_id = format!("scan_{}_{}", library_id, chrono::Utc::now().timestamp());
     let _ = job_system.app_handle().emit("job_started", &job_id);
 
-    let db = state.lock().map_err(|e| e.to_string())?;
-    let library_service = LibraryService::new(&db);
-    let library = library_service.get_library(library_id).map_err(|e| e.to_string())?;
-    let scanner = FileScanner::new(&db);
-    let result = scanner
-        .scan_library(library_id, &library.root_path)
-        .map_err(|e| e.to_string())?;
+    let (result, new_asset_ids) = {
+        let db = state.lock().map_err(|e| e.to_string())?;
+        let library_service = LibraryService::new(&db);
+        let library = library_service.get_library(library_id).map_err(|e| e.to_string())?;
+        let scanner = FileScanner::new(&db);
+        let result = scanner
+            .scan_library(library_id, &library.root_path)
+            .map_err(|e| e.to_string())?;
+
+        let conn = db.connection();
+        let mut stmt = conn
+            .prepare("SELECT id FROM assets WHERE library_id = ? AND thumb_status = 'none' ORDER BY id DESC LIMIT ?")
+            .map_err(|e| e.to_string())?;
+        let asset_ids: Vec<i64> = stmt
+            .query_map(rusqlite::params![library_id, result.added], |row| row.get(0))
+            .map_err(|e| e.to_string())?
+            .filter_map(|r| r.ok())
+            .collect();
+
+        (result, asset_ids)
+    };
+
+    for asset_id in new_asset_ids {
+        job_system.queue_thumbnail(asset_id);
+    }
 
     let _ = job_system.app_handle().emit("job_completed", &job_id);
