+++ src/features/libraries/AddLibraryDialog.tsx
@@ -1,6 +1,6 @@
 import { useState, type FormEvent, type ChangeEvent } from "react";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { addLibrary } from "@/shared/api/client";
+import { addLibrary, selectFolder } from "@/shared/api/client";
 import { useAppStore } from "@/shared/hooks/useAppStore";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
@@ -41,8 +41,10 @@ export function AddLibraryDialog({ open, onOpenChange }: AddLibraryDialogProps) 
   });
 
   const handleSelectFolder = async () => {
-    // TODO: Implement folder selection using @tauri-apps/plugin-dialog
-    // For now, user must manually enter the path
+    const selected = await selectFolder("Select Library Folder");
+    if (selected) {
+      setPath(selected);
+    }
   };
 
   const handleSubmit = (e: FormEvent) => {
