+++ src/shared/api/client.ts
@@ -1,4 +1,5 @@
 import { invoke } from "@tauri-apps/api/core";
+import { open } from "@tauri-apps/plugin-dialog";
 import type {
   Asset,
   Library,
@@ -66,6 +67,10 @@ export async function listAssets(query: AssetQuery = {}) {
 
 export async function getAssetDetail(assetId: number) {
   return invoke<Asset>("get_asset_detail", { assetId });
+}
+
+export async function getAssetNote(assetId: number) {
+  return invoke<string | null>("get_asset_note", { assetId });
 }
 
 export async function updateAssetNote(assetId: number, content: string) {
@@ -168,4 +173,16 @@ export async function moveAssets(
     destinationFolder,
     conflictPolicy,
   });
+}
+
+export async function selectFolder(title: string) {
+  const selected = await open({
+    directory: true,
+    multiple: false,
+    title,
+  });
+  if (selected && typeof selected === "string") {
+    return selected;
+  }
+  return null;
 }
